refactor: remove duplicated default file
